Decouple query support from validation base classes
