+++ src/MVVMFluent/ValidationViewModelBase.cs
@@ -1,7 +1,6 @@
-using MVVMFluent.Commands;
 using MVVMFluent.Interfaces;
-using MVVMFluent.Queries;
 using MVVMFluent.Validation;
+using MVVMFluent.Queries;
 using System;
 using System.Collections;
 using System.Collections.Generic;
@@ -13,13 +12,10 @@ using System.Runtime.CompilerServices;
 namespace MVVMFluent;
 
 /// <summary>
-/// Represents a base class for view models that provides property change notification, command creation, and validation.
+/// Represents a base class for view models that provides property change notification, command creation, validation, and query commands.
 /// </summary>
-public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IValidationFluentSetterViewModel
+public abstract class ValidationViewModelBase : QueryViewModelBase, IValidationFluentSetterViewModel
 {
-    private readonly Dictionary<string, object> _queryBuilderStore = new();
-    private IQueryDispatcher? _queryDispatcher;
-
     /// <summary>
     /// Gets a value indicating whether any properties in this view model have validation errors.
     /// </summary>
@@ -41,7 +37,7 @@ public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IVali
     /// </summary>
     protected ValidationViewModelBase()
     {
-        ErrorsChanged += ErrorsChangedHandler;
+        InitializeValidation();
     }
 
     /// <summary>
@@ -49,9 +45,14 @@ public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IVali
     /// </summary>
     /// <param name="queryDispatcher">The dispatcher responsible for executing queries.</param>
     protected ValidationViewModelBase(IQueryDispatcher queryDispatcher)
-        : this()
+        : base(queryDispatcher)
     {
-        UseQueries(queryDispatcher);
+        InitializeValidation();
+    }
+
+    private void InitializeValidation()
+    {
+        ErrorsChanged += ErrorsChangedHandler;
     }
 
     private void ErrorsChangedHandler(object? sender, DataErrorsChangedEventArgs e)
@@ -145,8 +146,7 @@ public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IVali
     /// <param name="propertyName">The name of the property. Automatically captured from the caller member name.</param>
     /// <returns>An <see cref="IValidationFluentSetter{TValue}"/> that allows configuring validation rules, change callbacks, and notifications before committing the value.</returns>
     /// <exception cref="ArgumentNullException">Thrown when the property name cannot be determined.</exception>
-
-    protected IValidationFluentSetter<TValue> When<TValue>(TValue value, [CallerMemberName] string? propertyName = null)
+    protected new IValidationFluentSetter<TValue> When<TValue>(TValue value, [CallerMemberName] string? propertyName = null)
     {
         if (propertyName == null)
         {
@@ -162,50 +162,6 @@ public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IVali
         return new ValidationFluentSetterBuilder<TValue>(value, this, propertyName);
     }
 
-    /// <summary>
-    /// Configures the view model to use the supplied <see cref="IQueryDispatcher"/>.
-    /// </summary>
-    /// <param name="queryDispatcher">The dispatcher responsible for executing queries.</param>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="queryDispatcher"/> is <see langword="null"/>.</exception>
-    protected void UseQueries(IQueryDispatcher queryDispatcher)
-    {
-        _queryDispatcher = queryDispatcher ?? throw new ArgumentNullException(nameof(queryDispatcher));
-    }
-
-    /// <summary>
-    /// Creates a query command builder that dispatches the supplied query when executed.
-    /// </summary>
-    /// <typeparam name="TResult">The type of result produced by the query.</typeparam>
-    /// <param name="factory">The factory used to create the query instance.</param>
-    /// <param name="propertyName">The property name used for caching. Automatically provided by the compiler.</param>
-    /// <returns>A <see cref="QueryAsyncCommandBuilder{TResult}"/> for configuring the command.</returns>
-    /// <exception cref="ArgumentNullException">Thrown when <paramref name="factory"/> or <paramref name="propertyName"/> is null.</exception>
-    /// <exception cref="InvalidOperationException">Thrown when no query dispatcher has been configured.</exception>
-    protected QueryAsyncCommandBuilder<TResult> Send<TResult>(Func<IQuery<TResult>> factory, [CallerMemberName] string? propertyName = null)
-    {
-        if (factory == null)
-        {
-            throw new ArgumentNullException(nameof(factory));
-        }
-
-        EnsurePropertyName(propertyName);
-
-        if (_queryDispatcher == null)
-        {
-            throw new InvalidOperationException("No IQueryDispatcher has been configured. Provide one via the constructor or call UseQueries before creating query commands.");
-        }
-
-        if (_queryBuilderStore.TryGetValue(propertyName!, out var existingBuilder))
-        {
-            return (QueryAsyncCommandBuilder<TResult>)existingBuilder;
-        }
-
-        var builder = new QueryAsyncCommandBuilder<TResult>(this, factory, _queryDispatcher);
-        _queryBuilderStore[propertyName!] = builder;
-        _commandStore[propertyName!] = builder.Command;
-        return builder;
-    }
-
     private IEnumerable GetAllErrors()
     {
         return _builderStore.Values
@@ -218,23 +174,21 @@ public abstract class ValidationViewModelBase : FluentSetterViewModelBase, IVali
     /// </summary>
     protected override void DisposeInternal()
     {
-        _queryBuilderStore.Clear();
         ErrorsChanged -= ErrorsChangedHandler;
         ErrorsChanged = null;
 
         base.DisposeInternal();
     }
 
-    private static void EnsurePropertyName(string? propertyName)
-    {
-        if (propertyName == null)
-        {
-            throw new ArgumentNullException(nameof(propertyName), "Not able to determine property name.");
-        }
+    /// <summary>
+    /// Gets a value that indicates whether the view model has any validation errors.
+    /// </summary>
+    bool INotifyDataErrorInfo.HasErrors => HasErrors;
 
-        if (propertyName == ".ctor")
-        {
-            throw new ArgumentException("Property name must be provided when it is used inside a constructor.", nameof(propertyName));
-        }
-    }
+    /// <summary>
+    /// Returns the validation errors for the specified property.
+    /// </summary>
+    /// <param name="propertyName">The name of the property whose errors are to be retrieved.</param>
+    /// <returns>An <see cref="IEnumerable"/> containing the validation errors for the specified property.</returns>
+    IEnumerable INotifyDataErrorInfo.GetErrors(string? propertyName) => GetErrors(propertyName);
 }
